+++ scripts/hooks.js
@@ -117,6 +117,7 @@ Hooks.on('renderTokenHUD', (tokenHUD, html, app) => {
       inputBox.change(async (inputbox) => {
         if (token === undefined) return;
         await engine.setHiddenValue(actor, duplicate(hiddenEffect), flag, Number(inputbox.target.value));
+        game.stealthy.socket.executeForEveryone('RefreshPerception');
       });
     }